Refactoriing
