+++ src/main/java/io/github/classgraph/utils/NestedJarHandler.java
@@ -314,8 +314,8 @@ public class NestedJarHandler {
                         // Extract nested jar to a temporary file
                         try {
                             File childJarFile = null;
-                            if (canUseSpringBootApi) {
-                                if (parentZipFile instanceof JarFile && childZipEntry.getMethod() == ZipEntry.STORED) {
+                            if (canUseSpringBootApi && childZipEntry.getMethod() == ZipEntry.STORED) {
+                                if (parentZipFile instanceof JarFile) {
                                     childJarFile = new NestedJarFile((JarFile) parentZipFile, childZipEntry);
                                 }
                             }
@@ -784,6 +784,60 @@ public class NestedJarHandler {
                             (success ? "Removed" : "Unable to remove") + " " + path + (e == null ? "" : " : " + e));
                 }
             }
+        }
+    }
+
+    /**
+     * Fake {@link File} for nested jar files handled by Spring-Boot's {@link JarFile}-API.
+     *
+     * This class is only loaded when the Spring-Boot API is available, so it can use the
+     * Spring-Boot types in it's interface.
+     */
+    static class NestedJarFile extends File {
+
+        private final JarFile parentJarFile;
+        private final ZipEntry nestedJarEntry;
+        private final boolean directory;
+
+        NestedJarFile(JarFile parentJarFile, ZipEntry nestedJarEntry) {
+            super(parentJarFile.getName() + "!/" + nestedJarEntry.getName());
+            this.parentJarFile = parentJarFile;
+            this.nestedJarEntry = nestedJarEntry;
+            this.directory = nestedJarEntry.isDirectory();
+        }
+
+        JarFile openAsJarFile() throws IOException {
+            return parentJarFile.getNestedJarFile(nestedJarEntry);
+        }
+
+        @Override
+        public boolean canRead() {
+            return true;
+        }
+
+        @Override
+        public boolean canWrite() {
+            return false;
+        }
+
+        @Override
+        public boolean isFile() {
+            return !directory;
+        }
+
+        @Override
+        public boolean isDirectory() {
+            return directory;
+        }
+
+        @Override
+        public boolean exists() {
+            return true;
+        }
+
+        @Override
+        public File getCanonicalFile() {
+            return this;
         }
     }
 }
